move cloud class to directory
